moving the example into a folder
